IMS_Backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.0,<24.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Environment
python-decouple>=3.8,<4.0

# CORS
django-cors-headers>=4.3,<5.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Utilities
django-extensions>=3.2,<4.0
